refactor: extract ConfigurationLoader from Program.cs
